Revert "Børre if this works I'm mad"
This reverts commit 971b1a723d75ef4b4063ad6f3c7bc7886138f20e.

This works, because voices are being filtered on gender and language. For the female voices, it leaves only one entry in the voices array. It is thus better to render the name as some kind of text element, instead of giving the illusion of having more than one voice with a one element long select.